requirements.txt
aiohttp[socks]>=3.8.0
httpx>=0.24.0
dnspython>=2.3.0
pyyaml>=6.0
paramiko>=2.11.0
pymysql>=1.0.2
psycopg2-binary>=2.9.5
redis>=4.5.0
pymongo>=4.3.0
Flask>=2.3.0
Flask-Login>=0.6.2
SQLAlchemy>=2.0.0
pyTelegramBotAPI>=4.0.0

